Expand telephone number

+++ pkg/client/model_j1_segment.go
@@ -22,7 +22,7 @@ type J1Segment struct {
 	GenerationCode               string    `json:"generationCode,omitempty"`
 	SocialSecurityNumber         int32     `json:"socialSecurityNumber"`
 	DateBirth                    time.Time `json:"dateBirth"`
-	TelephoneNumber              int32     `json:"telephoneNumber,omitempty"`
+	TelephoneNumber              int64     `json:"telephoneNumber,omitempty"`
 	EcoaCode                     string    `json:"ecoaCode"`
 	ConsumerInformationIndicator string    `json:"consumerInformationIndicator,omitempty"`
 }
